first commit
so i've gotten feedback from 2.0 and yeah the ui sucks. and so does the code. it's all a big jumbled mess and mobile is hot garbage so im just revamping everything visually. i'm starting with the upgrades panel, which is going to be combined with stats. here's my first draft